requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
google-cloud-bigquery>=3.10.0
python-dotenv>=1.0.0